handlers splitted and prework for GUI adding
