+++ data/20230825_100957_LCM_inlet=ring_outlet=middle/data.csv
@@ -1,0 +1,7 @@
+timestamp,Twarm,Tcold,Tout,h=35mm,h=65mm,h=95mm,h=125mm,h=155mm,h=185mm,h=215mm,warm_massflow,cold_massflow,warm/cold
+20230825_100954_816349,29.811,29.247,27.2985,26.841,26.7645,26.736,26.7195,26.8215,26.64,26.8455,0.0,0.0,False
+20230825_100955_325640,29.808,29.328,27.3075,26.826,26.7795,26.7345,26.718,26.916,26.637,26.8305,0.0,0.0,False
+20230825_100955_840619,29.754,29.328,27.3,26.7735,26.7825,26.6595,26.739,26.8185,26.652,26.847,0.0,0.0,False
+20230825_100956_349721,29.688,29.304,27.315,26.7645,26.7795,26.727,26.7375,26.8215,26.658,26.829,0.0,0.0,False
+20230825_100956_868042,29.6685,29.31,27.306,26.7735,26.763,26.6355,26.733,26.829,26.637,26.8215,0.0,0.0,False
+20230825_100957_384198,29.751,29.319,27.3255,26.847,26.7615,26.652,26.7315,26.8425,26.6565,26.8305,0.0,0.0,False
