+++ data/20230825_101043_LCM_inlet=ring_outlet=middle/Experiment_configuration.txt
@@ -1,0 +1,41 @@
++++++++++++++++++++++++++
+LCM EXPERIMENT PARAMETER 
++++++++++++++++++++++++++
+
+DATE AND TIME: 
+------------------------------
+Date: 25.August.2023
+Time: 10:10 Uhr
+
+SETTINGS: 
+------------------------------
+Inlet: single
+Outlet: middle
+Thermostat: 60
+
+COMMENT: 
+------------------------------
+A clever comment of the experiment I have made ;-)
+
+MEASUREMENT INFORMATION: 
+------------------------------
+Sample Rate: 3
+
+LABJACK INFORMATION: 
+------------------------------
+PT100:
+	 Twarm: AIN0
+	 Tcold: AIN1
+	 Tout: AIN2
+	 h=35mm: AIN3
+	 h=65mm: AIN4
+	 h=95mm: AIN5
+	 h=125mm: AIN6
+	 h=155mm: AIN7
+	 h=185mm: AIN8
+	 h=215mm: AIN9
+Flowmeter:
+	 warm_massflow: AIN11
+	 cold_massflow: AIN13
+Switch:
+	 warm/cold: AIN10
