+++ data/20230825_101043_LCM_inlet=ring_outlet=middle/data.csv
@@ -1,0 +1,18 @@
+timestamp,Twarm,Tcold,Tout,h=35mm,h=65mm,h=95mm,h=125mm,h=155mm,h=185mm,h=215mm,warm_massflow,cold_massflow,warm/cold
+20230825_101035_225851,29.6505,29.2995,27.303,26.7645,26.784,26.793,26.736,26.6745,26.6595,26.8995,0.0,0.0,False
+20230825_101035_739487,29.6745,29.229,27.3255,26.748,26.7825,26.796,26.739,26.67,26.571,26.8365,0.0,0.0,False
+20230825_101036_254134,29.7495,29.2335,27.315,26.7495,26.6895,26.7345,26.6655,26.7465,26.5605,26.907,0.0,0.0,False
+20230825_101036_772555,29.745,29.31,27.3105,26.7585,26.7045,26.64,26.727,26.742,26.643,26.8425,0.0,0.0,False
+20230825_101037_281984,29.727,29.328,27.3015,26.7495,26.7075,26.634,26.724,26.898,26.637,26.844,0.0,0.0,False
+20230825_101037_798657,29.754,29.319,27.3105,26.763,26.7855,26.724,26.742,26.844,26.565,26.8425,0.0,0.0,False
+20230825_101038_312987,29.7315,29.301,27.3015,26.751,26.76,26.733,26.7195,26.8155,26.58,26.832,0.0,0.0,False
+20230825_101038_843180,29.748,29.304,27.324,26.772,26.76,26.631,26.7195,26.763,26.646,26.8455,0.0,0.0,False
+20230825_101039_367180,29.7465,29.241,27.381,26.763,26.769,26.6505,26.664,26.682,26.652,26.832,0.0,0.0,False
+20230825_101039_872545,29.7525,29.301,27.3045,26.769,26.7585,26.64,26.7405,26.76,26.6415,26.919,0.0,0.0,False
+20230825_101040_374268,29.73,29.3265,27.327,26.7495,26.721,26.6415,26.691,26.742,26.6355,26.922,0.0,0.0,False
+20230825_101040_890441,29.7525,29.3145,27.3165,26.7585,26.6865,26.634,26.7195,26.742,26.559,26.8305,0.0,0.0,False
+20230825_101041_400977,29.754,29.328,27.309,26.7495,26.7705,26.7105,26.742,26.697,26.562,26.907,0.0,0.0,False
+20230825_101041_908530,29.739,29.403,27.303,26.7585,26.772,26.727,26.736,26.589,26.6355,26.829,0.0,0.0,False
+20230825_101042_413287,29.748,29.397,27.321,26.7495,26.772,26.733,26.7435,26.61,26.6385,26.8245,0.0,0.0,False
+20230825_101042_932990,29.7405,29.3295,27.3255,26.7585,26.775,26.727,26.7195,26.6865,26.64,26.8275,0.0,0.0,False
+20230825_101043_448966,29.73,29.31,27.3285,26.772,26.772,26.6355,26.727,26.76,26.6355,26.823,0.0,0.0,False
